Reorganized the document element code as a single component.
