+++ src/client/components/document-management-components.js
@@ -1,12 +1,28 @@
 import _ from 'lodash';
 import h from 'react-hyperscript';
 import React from 'react';
+import { Link } from 'react-router-dom';
 import { format, formatDistanceToNow, isThisMonth } from 'date-fns';
 
+import Document from '../../model/document';
 import {
-  tryPromise, sendMail, makeClassList } from '../../util';
+  tryPromise,
+  sendMail,
+  makeClassList
+} from '../../util';
 import logger from '../logger';
 import DirtyComponent from './dirty-component';
+import {
+  PUBMED_LINK_BASE_URL,
+  DOI_LINK_BASE_URL,
+  EMAIL_TYPE_INVITE,
+  EMAIL_TYPE_FOLLOWUP
+} from '../../config' ;
+
+const DOCUMENT_STATUS_FIELDS = Document.statusFields();
+
+const hasIssues = doc => _.values( doc.issues() ).some( i => !_.isNull( i ) );
+const hasIssue = ( doc, key ) => _.has( doc.issues(), key ) && !_.isNull( _.get( doc.issues(), key ) );
 
 class SendingComponent extends React.Component {
   constructor( props ){
@@ -216,9 +232,228 @@ class TextEditableComponent extends React.Component {
   }
 }
 
+class DocumentManagementDocumentComponent extends React.Component {
+  constructor( props ){
+    super( props );
+  }
+
+  render(){
+    let { doc, apiKey } = this.props;
+
+    // Document Header & Footer
+    const getDocumentHeader = doc => {
+      return h( 'div.document-management-document-section.meta', [
+        h( 'div.document-management-document-section-items', [
+          h( 'div', [
+            h( 'i.material-icons.hide-by-default.invalid', {
+              className: makeClassList({ 'show': hasIssues( doc ) })
+            }, 'warning' ),
+            h( 'i.material-icons.hide-by-default.mute', {
+              className: makeClassList({ 'show': doc.approved() })
+            }, 'thumb_up' ),
+            h( 'i.material-icons.hide-by-default.complete', {
+              className: makeClassList({ 'show': doc.submitted() })
+            }, 'check_circle' ),
+            h( 'i.material-icons.hide-by-default.mute', {
+              className: makeClassList({ 'show': doc.trashed() })
+            }, 'delete' )
+          ]),
+
+        ])
+      ]);
+    };
+
+    // Article
+    const getDocumentArticle = doc => {
+      let items = null;
+      // const { paperId: initial } = doc.provided();
+      // let labelContent = labelContent = h( TextEditableComponent, {
+      //   doc, initial, apiKey
+      // });
+
+      if( hasIssue( doc, 'paperId' ) ){
+        const { paperId: paperIdIssue } = doc.issues();
+        items = [
+          h( 'div', [
+            h( 'i.material-icons', 'error_outline' ),
+            h( 'span', ` ${paperIdIssue}` )
+          ])
+        ];
+
+      } else {
+        const { authors, contacts, title, reference, pmid, doi } = doc.citation();
+        const contactList = contacts.map( contact => `${contact.email} <${contact.name}>` ).join(', ');
+        items =  [
+            h( 'strong', [
+              h( 'a.plain-link.section-item-emphasize', {
+                href: PUBMED_LINK_BASE_URL + pmid,
+                target: '_blank'
+              }, title )
+            ]),
+            h('small.mute', `${authors}. ${reference}` ),
+            h( 'small.mute', [
+              h( 'a.plain-link', {
+                href: DOI_LINK_BASE_URL + doi,
+                target: '_blank'
+              }, `DOI: ${doi}` )
+            ]),
+            h('small.mute', contactList)
+          ];
+      }
+
+      return h( 'div.document-management-document-section', [
+        h( 'div.document-management-document-section-label', {
+          className: makeClassList({ 'issue': hasIssue( doc, 'paperId' ) })
+        }, [
+          h( 'div.document-management-document-section-label-text', 'Article:')
+        ]),
+        h( 'div.document-management-document-section-items', items )
+      ]);
+    };
+
+    // Document
+    const getDocumentInfo = doc => {
+      return h( 'div.document-management-document-section', [
+          h( 'div.document-management-document-section-label', [
+            h( 'div.document-management-document-section-label-text', 'Document:'),
+            h( 'div.document-management-document-section-label-content', [
+              h( DocumentRefreshButtonComponent, {
+                workingMessage: 'Refreshing',
+                disableWhen: doc.trashed(),
+                buttonKey: doc.id(),
+                params: { id: doc.id(), secret: doc.secret(), apiKey },
+                value: 'refresh',
+                title: 'Refresh document data',
+                label: h( 'i.material-icons', 'refresh' )
+              })
+            ])
+          ]),
+          h( 'div.document-management-document-section-items', [
+            h( 'div.document-management-document-section-items-row', [
+              h( Link, {
+                className: 'plain-link',
+                to: doc.publicUrl(),
+                target: '_blank',
+              }, 'Summary' ),
+              h( Link, {
+                className: 'plain-link',
+                to: doc.privateUrl(),
+                target: '_blank'
+              }, 'Editable' )
+            ])
+          ])
+        ]);
+    };
+
+    // Correspondence
+    const getContact = doc => {
+      const { authorEmail } = doc.correspondence();
+      const { contacts } = doc.citation();
+      return _.find( contacts, contact => _.indexOf( _.get( contact, 'email' ), authorEmail ) > -1 );
+    };
+
+    const getAuthorEmail = doc => {
+      const { authorEmail, isCorrespondingAuthor } = doc.correspondence();
+      let contact = getContact( doc );
+      const element = [ h( 'span', `${authorEmail} ` ) ];
+      if( contact ) element.push( h( 'span', ` <${contact.name}> ` ) );
+      if( isCorrespondingAuthor ) element.push( h( 'small', '(Corresponding)' ) );
+      return element;
+    };
+
+    const getDocumentCorrespondence = doc => {
+      let content = null;
+      if( hasIssue( doc, 'authorEmail' ) ){
+        const { authorEmail } = doc.issues();
+        content = h( 'div.document-management-document-section-items', [
+          h( 'div', [
+            h( 'i.material-icons', 'error_outline' ),
+            h( 'span', ` ${authorEmail}` )
+          ])
+        ]);
+
+      } else {
+        content = h( 'div.document-management-document-section-items', [
+          h( 'div', getAuthorEmail( doc ) ),
+          h( DocumentEmailButtonComponent, {
+            params: { doc, apiKey },
+            workingMessage: 'Sending...',
+            buttonKey: EMAIL_TYPE_INVITE,
+            value: EMAIL_TYPE_INVITE,
+            label: _.capitalize( EMAIL_TYPE_INVITE ),
+            disableWhen: doc.requested() || doc.trashed()
+          }),
+          h( DocumentEmailButtonComponent, {
+            params: { doc, apiKey },
+            workingMessage: 'Sending...',
+            buttonKey: EMAIL_TYPE_FOLLOWUP,
+            value: EMAIL_TYPE_FOLLOWUP,
+            label: _.upperFirst( EMAIL_TYPE_FOLLOWUP.replace(/([A-Z])/g, (match, letter) => '-' + letter) ),
+            disableWhen: doc.requested() || doc.trashed()
+          })
+        ]);
+      }
+
+      return h( 'div.document-management-document-section', [
+        h( 'div.document-management-document-section-label', {
+          className: makeClassList({ 'issue': hasIssue( doc, 'authorEmail' ) })
+        }, 'Correspondence:' ),
+        content
+      ]);
+    };
+
+    // Status
+    const getDocumentStatus = doc => {
+      let radios = [];
+      let addType = (typeVal, displayName) => {
+        radios.push(
+          h('input', {
+            type: 'radio',
+            name: `document-status-${doc.id()}`,
+            id: `document-status-radio-${doc.id()}-${typeVal}`,
+            value: typeVal,
+            defaultChecked: _.get( DOCUMENT_STATUS_FIELDS, typeVal ) === doc.status(),
+            onChange: e => {
+              let newlySelectedStatus = _.get( DOCUMENT_STATUS_FIELDS, e.target.value );
+              doc.status( newlySelectedStatus );
+            }
+          }),
+          h('label', {
+            htmlFor: `document-status-radio-${doc.id()}-${typeVal}`
+          }, displayName)
+        );
+      };
+
+      _.toPairs( DOCUMENT_STATUS_FIELDS ).forEach( ([ field, status ]) => addType( field, _.capitalize( status ) ) );
+      return h( 'div.radioset', radios );
+    };
+
+    // Stats
+    const getDocumentStats = doc => {
+      const created = toPeriodOrDate( doc.createdDate() );
+      const edited = toPeriodOrDate( doc.lastEditedDate() );
+      const context = doc.correspondence() ? _.get( doc.correspondence(), 'context' ) : null;
+      const source = context ? `via ${context}` : '';
+      return h( 'div.document-management-document-section.column.meta', [
+          h( 'small.document-management-document-section-items', [
+            getDocumentStatus( doc ),
+            h( 'div.mute', { key: 'created' }, `Created ${created} ${source}` ),
+            h( 'div.mute', { key: 'edited' }, edited ? `Edited ${edited}`: 'Not edited' )
+          ])
+        ]);
+    };
+
+    return h('div.document-management-document', [
+      getDocumentHeader( doc ),
+      getDocumentArticle( doc ),
+      getDocumentInfo( doc ),
+      getDocumentCorrespondence( doc ),
+      getDocumentStats( doc )
+    ]);
+  }
+}
+
 
 export {
-  DocumentEmailButtonComponent,
-  DocumentRefreshButtonComponent,
-  TextEditableComponent
+  DocumentManagementDocumentComponent
 };
